% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_typos.R
\name{generate_typos}
\alias{generate_typos}
\title{Generate typos using multiple methods}
\usage{
generate_typos(vec, methods, chars = NA)
}
\arguments{
\item{vec}{A vector of strings to attempt to generate typos for.}

\item{methods}{A vector of strings matching the methods (listed in
\link{typo_addition}) to use for typo generation.}

\item{chars}{Characters to use in generating typos. If set to \code{NA}, will use
the default character sets for each method; otherwise, \code{chars[[i]]} should be
a character set usable by \code{methods[[i]]}.}
}
\value{
A named list of length \code{length(vec)} containing all unique
permutations from the typo-generating methods.
}
\description{
This is a simple wrapper to let users generate typos using multiple methods
in a single function call. Results returned by this function should be
equivalent to the union of all functions supplied to \code{methods}.
}
\examples{
generate_typos(c("Michael", "David"), c("typo_addition", "typo_prefix"))

}
