% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charsets.R
\name{gh_allowed}
\alias{gh_allowed}
\title{Create a vector of characters allowed in GitHub repository names}
\usage{
gh_allowed()
}
\value{
A character vector of length 39 containing characters allowed in
GitHub repository names.
}
\description{
Create a vector of characters allowed in GitHub repository names
}
\examples{
gh_allowed()
}
