% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charsets.R
\name{qwerty_neighbors}
\alias{qwerty_neighbors}
\title{Create a list of characters which neighbor one another on QWERTY keyboards}
\usage{
qwerty_neighbors()
}
\value{
A list of length 37. Each element of the list is named with a key
on the QWERTY keyboard and contains a character vector of neighboring keys.
}
\description{
Create a list of characters which neighbor one another on QWERTY keyboards
}
\examples{
qwerty_neighbors()
}
